requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
distlib==0.3.8
fastapi==0.110.0
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
lxml==5.1.0
MarkupSafe==2.1.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
platformdirs==4.2.0
pydantic==2.6.2
pydantic_core==2.16.3
PyMySQL==1.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
reactpy==1.0.2
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
virtualenv==20.25.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
